feat: refactor for TUI multi panel
