+++ crates/jira/src/tui/mod.rs
@@ -1,3 +1,10 @@
-pub mod app;
+mod app;
+mod column;
+mod keys;
+mod mode;
+mod picker;
+mod prefs;
+mod prompts;
+mod render;
 
 pub use app::run_tui;
