+++ crates/jira/src/tui/mode.rs
@@ -1,0 +1,11 @@
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(super) enum Mode {
+    List,
+    View,
+    Search,
+    Transition,
+    Help,
+    ColumnPicker,
+    AssigneePicker,
+    ComponentPicker,
+}
